add _mixins suffix back to workaround some code completion bugs
