+++ packages/api-server/api_server/models/mixins/rmf_dispenser_msgs_mixins.py
@@ -5,7 +5,7 @@
 ########################################################
 
 from tortoise import fields
-from ._ros_time import RosTimeField
+from .ros_time import RosTimeField
 
 
 class DispenserRequestMixin():
